% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced.R
\name{gg_sced}
\alias{gg_sced}
\title{gg_sced}
\usage{
gg_sced(plt, legs, offs = NULL, verbose = FALSE)
}
\arguments{
\item{plt}{ggplot object as typically designed/printed in userspace}

\item{legs}{list of 'legs' to be drawn}

\item{offs}{TODO mapping of lines that require minor offset}
}
\value{
Finalized figure with respective phase change lines embedded.
}
\description{
Core exported function to facilitate the drawing of phase change lines atop a ggplot object. Primarily designed to be used *after* the plot is finalized, with the lines being the last element drawn at the highest z index (i.e., atop all elements).
}
