% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_reporting.R
\name{gg_sced_output_console}
\alias{gg_sced_output_console}
\title{gg_sced_output_console}
\usage{
gg_sced_output_console(msg, output)
}
\arguments{
\item{msg}{Message to be output to console}

\item{output}{Boolean to discern whether or not to output message}
}
\description{
Optional output function to support debugging (mainly for testing)
}
