% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced.R
\name{gg_sced_style_x}
\alias{gg_sced_style_x}
\title{gg_sced_style_x}
\usage{
gg_sced_style_x(expansion = 0, lwd = 2, col = "black")
}
\arguments{
\item{expansion}{Percentage of axis to pad (top/bottom) to break axis}

\item{lwd}{Width of axis line (clipping on)}

\item{col}{Color of axis drawn (default = 'black')}
}
\value{
Annotation to simulate a disconnected y-axis via expansion
}
\description{
gg_sced_style_x
}
