% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_helpers.R
\name{ggsced_condition_labels}
\alias{ggsced_condition_labels}
\title{generate_condition_labels}
\usage{
ggsced_condition_labels(plt, y = NULL)
}
\arguments{
\item{plt}{ggplot object}

\item{y}{ordinate of respective label (Default = last tick from plot object)}
}
\value{
a data frame that can be used with a geom_text layer
}
\description{
Optional helper class to prepare condition labels (e.g., Baseline, Intervention) based on the grouping of an existing ggplot object
}
