% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_utils.R
\name{ggsced_extract_domain}
\alias{ggsced_extract_domain}
\title{ggsced_extract_domain}
\usage{
ggsced_extract_domain(panel_params)
}
\arguments{
\item{panel_params}{respective parameters of panel grob}
}
\value{
numeric vector specifying size in min-max units
}
\description{
Internal helper to pull respective panel parameters and assert validity.
}
