% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_helpers.R
\name{ggsced_facet_labels}
\alias{ggsced_facet_labels}
\title{ggsced_facet_labels}
\usage{
ggsced_facet_labels(plt, y = 0, x = NULL)
}
\arguments{
\item{plt}{ggplot object}

\item{y}{ordinate of respective label (typically very low or very high)}

\item{x}{position on the x-axis (Default = last tick from plot object)}
}
\value{
a data frame that can be used with a geom_text layer
}
\description{
Optional helper class to prepare facet labels (e.g., Participant names) based on the faceting of an existing ggplot object
}
