% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_utils.R
\name{ggsced_get_panels}
\alias{ggsced_get_panels}
\title{ggsced_get_panels}
\usage{
ggsced_get_panels(ggplot_grobs)
}
\arguments{
\item{ggplot_grobs}{list of grobs prepared by ggplotGrob}
}
\value{
filtered list of grobs restricted to faceted levels
}
\description{
Honestly, a bit of a 'hackish' solution to extract grobs that are associated with data to be drawn. Probably fragile and specific to modern gg implementations, but probably the close we'll get. Subject to change.
}
