% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_theme.R
\name{ggsced_internal_x_axis}
\alias{ggsced_internal_x_axis}
\title{ggsced_internal_x_axis}
\usage{
ggsced_internal_x_axis(expansion = 0, lwd = 2, col = "black")
}
\arguments{
\item{expansion}{Percentage of axis to pad (left/right) to break axis (Note: normalized parent units)}

\item{lwd}{Width of axis line}

\item{col}{Color of axis drawn (default = 'black')}
}
\value{
Rendering of customized x axis across facets
}
\description{
Internal helper class to draw a conventional x-axis per SCED publication suggestions.
}
