% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_grobs.R
\name{ggsced_internal_y_axis}
\alias{ggsced_internal_y_axis}
\title{ggsced_internal_y_axis}
\usage{
ggsced_internal_y_axis(expansion = 0, lwd = 2, col = "black")
}
\arguments{
\item{expansion}{Percentage of axis to pad (top/bottom) to break axis (Note: normalized parent units)}

\item{lwd}{Width of axis line}

\item{col}{Color of axis drawn (default = 'black')}
}
\value{
Rendering of customized y axis across facets
}
\description{
Internal helper class to draw a conventional y-axis per SCED publication suggestions.
}
