% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_utils.R
\name{ggsced_name_dogleg_lateral}
\alias{ggsced_name_dogleg_lateral}
\title{ggsced_name_dogleg_lateral}
\usage{
ggsced_name_dogleg_lateral(panel, index, n_leg)
}
\arguments{
\item{panel}{name of the parent grob}

\item{index}{index of the phase change in sequence}

\item{n_leg}{leg in the phase change}
}
\value{
a name corresponding with unique phase-change grob element
}
\description{
Grobs need to have unique names lest they get accidentally overwritten. This is a convenience naming function for the 'lateral' part of a phase break within a facet level.
}
