% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_helpers.R
\name{ggsced_prep_labels}
\alias{ggsced_prep_labels}
\title{ggsced_prep_labels}
\usage{
ggsced_prep_labels(
  data,
  facet_col,
  specs,
  x_col = "Session",
  y_col = "Value",
  default_hadj = 0.5,
  default_vadj = 0.5,
  default_text_size = NA_real_,
  default_font_face = "plain"
)
}
\arguments{
\item{data}{A data frame containing the data to be used for plotting.}

\item{facet_col}{The name of the column in `data` that contains the facet values.}

\item{specs}{A named list where each name corresponds to a facet value and each entry is a list of label specifications. Each label specification should include at least the `label`, `x`, and `y` values, and can optionally include other aesthetics like `hadj`, `vadj`, `text_size`, and `fontface`.}

\item{x_col}{The name of the column in `data` that contains the x-axis values (default is "Session").}

\item{y_col}{The name of the column in `data` that contains the y-axis values (default is "Value").}

\item{default_hadj}{The default horizontal adjustment to use for labels if not specified in the `specs` (default is 0.5).}

\item{default_vadj}{The default vertical adjustment to use for labels if not specified in the `specs` (default is 0.5).}

\item{default_text_size}{The default text size to use for labels if not specified in the `specs` (default is NA_real_).}

\item{default_font_face}{The default font face to use for labels if not specified in the `specs` (default is "plain").}
}
\value{
A data frame containing the prepared label specifications for plotting, with additional columns for any extra aesthetics specified in the `specs`.
}
\description{
This function prepares a data frame for plotting labels in a ggplot object
based on specified facets and label specifications. It validates the input
parameters, checks for the presence of required columns, and constructs a new
data frame containing the necessary information for plotting labels according
to the provided specifications.
}
