% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_helpers.R
\name{ggsced_prep_lines}
\alias{ggsced_prep_lines}
\title{ggsced_prep_lines}
\usage{
ggsced_prep_lines(
  data,
  facet_col,
  specs,
  x_col = "Session",
  default_linetype = 1
)
}
\arguments{
\item{data}{A data frame containing the data to be used for plotting.}

\item{facet_col}{The name of the column in `data` that contains the facet values.}

\item{specs}{A named list where each name corresponds to a facet value and each entry is a list of line specifications. Each line specification should include at least the `x` value and can optionally include other aesthetics like `linetype`.}

\item{x_col}{The name of the column in `data` that contains the x-axis values (default is "Session").}

\item{default_linetype}{The default linetype to use for lines if not specified in the `specs` (default is 1).}
}
\value{
A data frame containing the prepared line specifications for plotting, with additional columns for any extra aesthetics specified in the `specs`.
}
\description{
This function prepares a data frame for plotting lines in a ggplot object
based on specified facets and line specifications. It validates the input
parameters, checks for the presence of required columns, and constructs a new
data frame containing the necessary information for plotting lines according
to the provided specifications.
}
