% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_utils.R
\name{ggsced_scale_units}
\alias{ggsced_scale_units}
\title{ggsced_scale_units}
\usage{
ggsced_scale_units(session_value, domain_size)
}
\arguments{
\item{session_value}{Session associated with phase-change line}

\item{domain_size}{Total size of domain per panel specification}
}
\value{
A proportional value given value in a panel domain
}
\description{
Internal helper to discern exactly where a vertical would need to be placed within a specific grob (i.e., panel grob). Uses the domain and phase change location as a reference to convert to npc units for gtable and grid.
}
