% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_theme.R
\name{ggsced_style_x}
\alias{ggsced_style_x}
\title{ggsced_style_x}
\usage{
ggsced_style_x(expansion = 0, lwd = 2, col = "black")
}
\arguments{
\item{expansion}{Percentage of axis to pad (left/right) to break axis (Note: normalized parent units)}

\item{lwd}{Width of axis line}

\item{col}{Color of axis drawn (default = 'black')}
}
\value{
Annotation to simulate a disconnected y-axis via expansion
}
\description{
This is a convenience element designed to support a more complete interpretation of SCED-specific style conventions through ggplot. Specifically, an axis is drawn with a specific manner of expansion to support the broken axis (i.e., non-connected x/y) expected of SCED-specific publication guidelines.
}
