% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LozyEtAl2020Data}
\alias{LozyEtAl2020Data}
\title{Plotting data from Lozy et al. (2020)}
\format{
A data frame with 91 rows and 5 variables:
\describe{
  \item{Session}{Session number}
  \item{Participant}{Participant name}
  \item{Phase}{Phase number}
  \item{Response}{Response type}
  \item{Value}{Response value}
}
}
\source{
<doi:https://doi.org/10.1002/jaba.702>
}
\usage{
LozyEtAl2020Data
}
\description{
Treatment data from Lozy et al. (2020)
}
\keyword{datasets}
