% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced_grobs.R
\name{sced_phase_change_simple_lateral_grob}
\alias{sced_phase_change_simple_lateral_grob}
\title{sced_phase_change_simple_lateral_grob}
\usage{
sced_phase_change_simple_lateral_grob(x0, x1, segment_name)
}
\arguments{
\item{x0}{value corresponding to respective domain in npc units}

\item{x1}{value corresponding to respective domain in npc units}

\item{segment_name}{name associated with object}
}
\value{
graphical object
}
\description{
This is a helper class to standardize drawing grobs
}
