% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsced.R
\name{gg_sced_scale_units}
\alias{gg_sced_scale_units}
\title{gg_sced_scale_units}
\usage{
gg_sced_scale_units(session_value, domain_size)
}
\arguments{
\item{session_value}{Session associated with phase-change line}

\item{domain_size}{Total size of domain per panel specification}
}
\value{
A proportional value given value in a panel domain
}
\description{
gg_sced_scale_units
}
\examples{
# Session number
session_num = 2

# Domain size
domain_size = c(1, 20)

gg_sced_scale_units(session_num, domain_size)

}
